+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 #### exonum
 
+- Storage in exonum is now versioned. Old databases will not work with this
+  update. (#707)
+
 - `Iron` based web API has been replaced by the new implementation based
   on `actix-web`. (#727)